+++ src/database/stream/query.rs
@@ -1,12 +1,9 @@
 #![allow(missing_docs, unreachable_code, unused_variables)]
 
 use tracing::instrument;
-
-#[cfg(any(feature = "mock", feature = "proxy"))]
-use std::sync::Arc;
 use std::{pin::Pin, task::Poll};
-
 use futures::Stream;
+
 #[cfg(feature = "sqlx-dep")]
 use futures::TryStreamExt;
 
@@ -29,101 +26,6 @@ pub struct QueryStream {
     stream: MetricStream<'this>,
 }
 
-#[cfg(feature = "sqlx-mysql")]
-impl
-    From<(
-        PoolConnection<sqlx::MySql>,
-        Statement,
-        Option<crate::metric::Callback>,
-    )> for QueryStream
-{
-    fn from(
-        (conn, stmt, metric_callback): (
-            PoolConnection<sqlx::MySql>,
-            Statement,
-            Option<crate::metric::Callback>,
-        ),
-    ) -> Self {
-        QueryStream::build(stmt, InnerConnection::MySql(conn), metric_callback)
-    }
-}
-
-#[cfg(feature = "sqlx-postgres")]
-impl
-    From<(
-        PoolConnection<sqlx::Postgres>,
-        Statement,
-        Option<crate::metric::Callback>,
-    )> for QueryStream
-{
-    fn from(
-        (conn, stmt, metric_callback): (
-            PoolConnection<sqlx::Postgres>,
-            Statement,
-            Option<crate::metric::Callback>,
-        ),
-    ) -> Self {
-        QueryStream::build(stmt, InnerConnection::Postgres(conn), metric_callback)
-    }
-}
-
-#[cfg(feature = "sqlx-sqlite")]
-impl
-    From<(
-        PoolConnection<sqlx::Sqlite>,
-        Statement,
-        Option<crate::metric::Callback>,
-    )> for QueryStream
-{
-    fn from(
-        (conn, stmt, metric_callback): (
-            PoolConnection<sqlx::Sqlite>,
-            Statement,
-            Option<crate::metric::Callback>,
-        ),
-    ) -> Self {
-        QueryStream::build(stmt, InnerConnection::Sqlite(conn), metric_callback)
-    }
-}
-
-#[cfg(feature = "mock")]
-impl
-    From<(
-        Arc<crate::MockDatabaseConnection>,
-        Statement,
-        Option<crate::metric::Callback>,
-    )> for QueryStream
-{
-    fn from(
-        (conn, stmt, metric_callback): (
-            Arc<crate::MockDatabaseConnection>,
-            Statement,
-            Option<crate::metric::Callback>,
-        ),
-    ) -> Self {
-        QueryStream::build(stmt, InnerConnection::Mock(conn), metric_callback)
-    }
-}
-
-#[cfg(feature = "proxy")]
-impl
-    From<(
-        Arc<crate::ProxyDatabaseConnection>,
-        Statement,
-        Option<crate::metric::Callback>,
-    )> for QueryStream
-{
-    fn from(
-        (conn, stmt, metric_callback): (
-            Arc<crate::ProxyDatabaseConnection>,
-            Statement,
-            Option<crate::metric::Callback>,
-        ),
-    ) -> Self {
-        QueryStream::build(stmt, InnerConnection::Proxy(conn), metric_callback)
-    }
-}
-
 impl std::fmt::Debug for QueryStream {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "QueryStream")
@@ -132,7 +34,7 @@ impl std::fmt::Debug for QueryStream {
 
 impl QueryStream {
     #[instrument(level = "trace", skip(metric_callback))]
-    fn build(
+    pub(crate) fn build(
         stmt: Statement,
         conn: InnerConnection,
         metric_callback: Option<crate::metric::Callback>,
