+++ src/driver/sqlx_mysql.rs
@@ -1,3 +1,4 @@
+use futures::lock::Mutex;
 use log::LevelFilter;
 use sea_query::Values;
 use std::{future::Future, pin::Pin, sync::Arc};
@@ -60,7 +61,7 @@ impl SqlxMySqlConnector {
                 );
             }
         }
-        match options.pool_options().connect_with(opt).await {
+        match options.sqlx_pool_options().connect_with(opt).await {
             Ok(pool) => Ok(DatabaseConnection::SqlxMySqlPoolConnection(
                 SqlxMySqlPoolConnection {
                     pool,
@@ -89,7 +90,7 @@ impl SqlxMySqlPoolConnection {
         debug_print!("{}", stmt);
 
         let query = sqlx_query(&stmt);
-        let mut conn = self.pool.acquire().await.map_err(conn_acquire_err)?;
+        let mut conn = self.pool.acquire().await.map_err(sqlx_conn_acquire_err)?;
         crate::metric::metric!(self.metric_callback, &stmt, {
             match query.execute(&mut *conn).await {
                 Ok(res) => Ok(res.into()),
@@ -103,7 +104,7 @@ impl SqlxMySqlPoolConnection {
     pub async fn execute_unprepared(&self, sql: &str) -> Result<ExecResult, DbErr> {
         debug_print!("{}", sql);
 
-        let conn = &mut self.pool.acquire().await.map_err(conn_acquire_err)?;
+        let conn = &mut self.pool.acquire().await.map_err(sqlx_conn_acquire_err)?;
         match conn.execute(sql).await {
             Ok(res) => Ok(res.into()),
             Err(err) => Err(sqlx_error_to_exec_err(err)),
@@ -116,7 +117,7 @@ impl SqlxMySqlPoolConnection {
         debug_print!("{}", stmt);
 
         let query = sqlx_query(&stmt);
-        let mut conn = self.pool.acquire().await.map_err(conn_acquire_err)?;
+        let mut conn = self.pool.acquire().await.map_err(sqlx_conn_acquire_err)?;
         crate::metric::metric!(self.metric_callback, &stmt, {
             match query.fetch_one(&mut *conn).await {
                 Ok(row) => Ok(Some(row.into())),
@@ -134,7 +135,7 @@ impl SqlxMySqlPoolConnection {
         debug_print!("{}", stmt);
 
         let query = sqlx_query(&stmt);
-        let mut conn = self.pool.acquire().await.map_err(conn_acquire_err)?;
+        let mut conn = self.pool.acquire().await.map_err(sqlx_conn_acquire_err)?;
         crate::metric::metric!(self.metric_callback, &stmt, {
             match query.fetch_all(&mut *conn).await {
                 Ok(rows) => Ok(rows.into_iter().map(|r| r.into()).collect()),
@@ -148,7 +149,7 @@ impl SqlxMySqlPoolConnection {
     pub async fn stream(&self, stmt: Statement) -> Result<QueryStream, DbErr> {
         debug_print!("{}", stmt);
 
-        let conn = self.pool.acquire().await.map_err(conn_acquire_err)?;
+        let conn = self.pool.acquire().await.map_err(sqlx_conn_acquire_err)?;
         Ok(QueryStream::from((
             conn,
             stmt,
@@ -163,7 +164,7 @@ impl SqlxMySqlPoolConnection {
         isolation_level: Option<IsolationLevel>,
         access_mode: Option<AccessMode>,
     ) -> Result<DatabaseTransaction, DbErr> {
-        let conn = self.pool.acquire().await.map_err(conn_acquire_err)?;
+        let conn = self.pool.acquire().await.map_err(sqlx_conn_acquire_err)?;
         DatabaseTransaction::new_mysql(
             conn,
             self.metric_callback.clone(),
@@ -189,7 +190,7 @@ impl SqlxMySqlPoolConnection {
         T: Send,
         E: std::error::Error + Send,
     {
-        let conn = self.pool.acquire().await.map_err(conn_acquire_err)?;
+        let conn = self.pool.acquire().await.map_err(sqlx_conn_acquire_err)?;
         let transaction = DatabaseTransaction::new_mysql(
             conn,
             self.metric_callback.clone(),
@@ -210,7 +211,7 @@ impl SqlxMySqlPoolConnection {
 
     /// Checks if a connection to the database is still valid.
     pub async fn ping(&self) -> Result<(), DbErr> {
-        let conn = &mut self.pool.acquire().await.map_err(conn_acquire_err)?;
+        let conn = &mut self.pool.acquire().await.map_err(sqlx_conn_acquire_err)?;
         match conn.ping().await {
             Ok(_) => Ok(()),
             Err(err) => Err(sqlx_error_to_conn_err(err)),
@@ -272,4 +273,180 @@ pub(crate) async fn set_transaction_config(
         conn.execute(query).await.map_err(sqlx_error_to_exec_err)?;
     }
     Ok(())
+}
+
+impl
+    From<(
+        PoolConnection<sqlx::MySql>,
+        Statement,
+        Option<crate::metric::Callback>,
+    )> for crate::QueryStream
+{
+    fn from(
+        (conn, stmt, metric_callback): (
+            PoolConnection<sqlx::MySql>,
+            Statement,
+            Option<crate::metric::Callback>,
+        ),
+    ) -> Self {
+        crate::QueryStream::build(stmt, crate::InnerConnection::MySql(conn), metric_callback)
+    }
+}
+
+impl crate::DatabaseTransaction {
+    pub(crate) async fn new_mysql(
+        inner: PoolConnection<sqlx::MySql>,
+        metric_callback: Option<crate::metric::Callback>,
+        isolation_level: Option<IsolationLevel>,
+        access_mode: Option<AccessMode>,
+    ) -> Result<crate::DatabaseTransaction, DbErr> {
+        Self::begin(
+            Arc::new(Mutex::new(crate::InnerConnection::MySql(inner))),
+            crate::DbBackend::MySql,
+            metric_callback,
+            isolation_level,
+            access_mode,
+        )
+        .await
+    }
+}
+
+#[cfg(feature = "proxy")]
+pub(crate) fn from_sqlx_mysql_row_to_proxy_row(row: &sqlx::mysql::MySqlRow) -> crate::ProxyRow {
+    // https://docs.rs/sqlx-mysql/0.7.2/src/sqlx_mysql/protocol/text/column.rs.html
+    // https://docs.rs/sqlx-mysql/0.7.2/sqlx_mysql/types/index.html
+    use sea_query::Value;
+    use sqlx::{Column, Row, TypeInfo};
+    crate::ProxyRow {
+        values: row
+            .columns()
+            .iter()
+            .map(|c| {
+                (
+                    c.name().to_string(),
+                    match c.type_info().name() {
+                        "TINYINT(1)" | "BOOLEAN" => Value::Bool(Some(
+                            row.try_get(c.ordinal()).expect("Failed to get boolean"),
+                        )),
+                        "TINYINT UNSIGNED" => Value::TinyUnsigned(Some(
+                            row.try_get(c.ordinal())
+                                .expect("Failed to get unsigned tiny integer"),
+                        )),
+                        "SMALLINT UNSIGNED" => Value::SmallUnsigned(Some(
+                            row.try_get(c.ordinal())
+                                .expect("Failed to get unsigned small integer"),
+                        )),
+                        "INT UNSIGNED" => Value::Unsigned(Some(
+                            row.try_get(c.ordinal())
+                                .expect("Failed to get unsigned integer"),
+                        )),
+                        "MEDIUMINT UNSIGNED" | "BIGINT UNSIGNED" => Value::BigUnsigned(Some(
+                            row.try_get(c.ordinal())
+                                .expect("Failed to get unsigned big integer"),
+                        )),
+                        "TINYINT" => Value::TinyInt(Some(
+                            row.try_get(c.ordinal())
+                                .expect("Failed to get tiny integer"),
+                        )),
+                        "SMALLINT" => Value::SmallInt(Some(
+                            row.try_get(c.ordinal())
+                                .expect("Failed to get small integer"),
+                        )),
+                        "INT" => Value::Int(Some(
+                            row.try_get(c.ordinal()).expect("Failed to get integer"),
+                        )),
+                        "MEDIUMINT" | "BIGINT" => Value::BigInt(Some(
+                            row.try_get(c.ordinal()).expect("Failed to get big integer"),
+                        )),
+                        "FLOAT" => Value::Float(Some(
+                            row.try_get(c.ordinal()).expect("Failed to get float"),
+                        )),
+                        "DOUBLE" => Value::Double(Some(
+                            row.try_get(c.ordinal()).expect("Failed to get double"),
+                        )),
+
+                        "BIT" | "BINARY" | "VARBINARY" | "TINYBLOB" | "BLOB" | "MEDIUMBLOB"
+                        | "LONGBLOB" => Value::Bytes(Some(Box::new(
+                            row.try_get(c.ordinal()).expect("Failed to get bytes"),
+                        ))),
+
+                        "CHAR" | "VARCHAR" | "TINYTEXT" | "TEXT" | "MEDIUMTEXT" | "LONGTEXT" => {
+                            Value::String(Some(Box::new(
+                                row.try_get(c.ordinal()).expect("Failed to get string"),
+                            )))
+                        }
+
+                        #[cfg(feature = "with-chrono")]
+                        "TIMESTAMP" => Value::ChronoDateTimeUtc(Some(Box::new(
+                            row.try_get(c.ordinal()).expect("Failed to get timestamp"),
+                        ))),
+                        #[cfg(all(feature = "with-time", not(feature = "with-chrono")))]
+                        "TIMESTAMP" => Value::TimeDateTime(Some(Box::new(
+                            row.try_get(c.ordinal()).expect("Failed to get timestamp"),
+                        ))),
+
+                        #[cfg(feature = "with-chrono")]
+                        "DATE" => Value::ChronoDate(Some(Box::new(
+                            row.try_get(c.ordinal()).expect("Failed to get date"),
+                        ))),
+                        #[cfg(all(feature = "with-time", not(feature = "with-chrono")))]
+                        "DATE" => Value::TimeDate(Some(Box::new(
+                            row.try_get(c.ordinal()).expect("Failed to get date"),
+                        ))),
+
+                        #[cfg(feature = "with-chrono")]
+                        "TIME" => Value::ChronoTime(Some(Box::new(
+                            row.try_get(c.ordinal()).expect("Failed to get time"),
+                        ))),
+                        #[cfg(all(feature = "with-time", not(feature = "with-chrono")))]
+                        "TIME" => Value::TimeTime(Some(Box::new(
+                            row.try_get(c.ordinal()).expect("Failed to get time"),
+                        ))),
+
+                        #[cfg(feature = "with-chrono")]
+                        "DATETIME" => Value::ChronoDateTime(Some(Box::new(
+                            row.try_get(c.ordinal()).expect("Failed to get datetime"),
+                        ))),
+                        #[cfg(all(feature = "with-time", not(feature = "with-chrono")))]
+                        "DATETIME" => Value::TimeDateTime(Some(Box::new(
+                            row.try_get(c.ordinal()).expect("Failed to get datetime"),
+                        ))),
+
+                        #[cfg(feature = "with-chrono")]
+                        "YEAR" => Value::ChronoDate(Some(Box::new(
+                            row.try_get(c.ordinal()).expect("Failed to get year"),
+                        ))),
+                        #[cfg(all(feature = "with-time", not(feature = "with-chrono")))]
+                        "YEAR" => Value::TimeDate(Some(Box::new(
+                            row.try_get(c.ordinal()).expect("Failed to get year"),
+                        ))),
+
+                        "ENUM" | "SET" | "GEOMETRY" => Value::String(Some(Box::new(
+                            row.try_get(c.ordinal())
+                                .expect("Failed to get serialized string"),
+                        ))),
+
+                        #[cfg(feature = "with-bigdecimal")]
+                        "DECIMAL" => Value::BigDecimal(Some(Box::new(
+                            row.try_get(c.ordinal()).expect("Failed to get decimal"),
+                        ))),
+                        #[cfg(all(
+                            feature = "with-rust_decimal",
+                            not(feature = "with-bigdecimal")
+                        ))]
+                        "DECIMAL" => Value::Decimal(Some(Box::new(
+                            row.try_get(c.ordinal()).expect("Failed to get decimal"),
+                        ))),
+
+                        #[cfg(feature = "with-json")]
+                        "JSON" => Value::Json(Some(Box::new(
+                            row.try_get(c.ordinal()).expect("Failed to get json"),
+                        ))),
+
+                        _ => unreachable!("Unknown column type: {}", c.type_info().name()),
+                    },
+                )
+            })
+            .collect(),
+    }
 }
