+++ src/error.rs
@@ -151,16 +151,6 @@ where
     DbErr::Json(s.to_string())
 }
 
-#[allow(dead_code)]
-#[cfg(feature = "sqlx-dep")]
-pub(crate) fn conn_acquire_err(sqlx_err: sqlx::Error) -> DbErr {
-    match sqlx_err {
-        sqlx::Error::PoolTimedOut => DbErr::ConnectionAcquire(ConnAcquireErr::Timeout),
-        sqlx::Error::PoolClosed => DbErr::ConnectionAcquire(ConnAcquireErr::ConnectionClosed),
-        _ => DbErr::Conn(RuntimeErr::SqlxError(sqlx_err)),
-    }
-}
-
 /// An error from unsuccessful SQL query
 #[derive(Error, Debug, Clone, PartialEq, Eq)]
 #[non_exhaustive]
